requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
graphql-core==3.2.6
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.265.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zeroc-ice==3.7.10.1